+++ library/forms/stub/src/stub_treenode.cpp
@@ -1,5 +1,5 @@
 /* 
- * Copyright (c) 2012, 2014, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2015, Oracle and/or its affiliates. All rights reserved.
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License as
@@ -186,7 +186,7 @@ void TreeNodeWrapper::move_child(mforms::TreeNodeRef node, int new_index)
   if (i == _children.end())
     return;
 
-  int old_index = i - _children.begin();
+  size_t old_index = i - _children.begin();
   if (old_index == new_index)
     return;
